List mani (#15)
* 14.1 Implemented
The random.txt file is inside bin/Debug

* Deletion
The directory Youngjin created because I still cannot access the designer.

* fix of filepath for random.txt
