--- hw06/List Manipulator/List Manipulator.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.7.34031.279
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "List Manipulator", "List Manipulator.csproj", "{3080647E-82D4-40FD-B8C4-10A3DD1FCCE3}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3080647E-82D4-40FD-B8C4-10A3DD1FCCE3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3080647E-82D4-40FD-B8C4-10A3DD1FCCE3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3080647E-82D4-40FD-B8C4-10A3DD1FCCE3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3080647E-82D4-40FD-B8C4-10A3DD1FCCE3}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {E6AC5EBF-3E77-42D4-91EF-584C1AC4F3AD}
-	EndGlobalSection
-EndGlobal
