+++ hw06/ListManipulator 14.1/random.txt
@@ -1,0 +1,100 @@
+88
+-62
+25
+93
+93
+57
+97
+19
+8
+-23
+-91
+-50
+45
+61
+26
+-5
+-37
+-8
+41
+88
+35
+23
+-85
+-95
+-59
+23
+-89
+-19
+-1
+-70
+58
+85
+17
+96
+-68
+-54
+24
+-67
+-6
+78
+85
+-99
+40
+-71
+72
+14
+91
+86
+26
+-30
+6
+39
+-59
+75
+15
+-35
+-77
+19
+52
+-31
+46
+15
+-81
+33
+18
+-76
+32
+49
+7
+-21
+70
+-70
+97
+62
+-98
+34
+82
+86
+-68
+26
+32
+-75
+55
+-71
+24
+-81
+44
+41
+-19
+67
+-77
+-96
+-19
+-75
+-68
+-30
+81
+-94
+11
+2
